# cmd/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Noble blockchain indexer",
        "title": "Noble Indexer API",
        "termsOfService": "https://bakingbad.dev/terms",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/baking-bad/noble-indexer",
            "email": "hello@bakingbad.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/baking-bad/noble-indexer/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "noble.dipdup.net",
    "basePath": "/v1",
    "paths": {
...
        "/block": {
            "get": {
                "description": "Returns a paginated list of blocks. Blocks can be sorted by height in ascending or descending order. Optionally includes statistical data for each block.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "List blocks",
                "operationId": "list-block",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of blocks to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of blocks to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order by height (default: asc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include statistics for each block (default: false)",
                        "name": "stats",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of blocks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Block"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/block/count": {
            "get": {
                "description": "Returns the total number of blocks indexed in the blockchain, including the genesis block",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Get total block count",
                "operationId": "get-blocks-count",
                "responses": {
                    "200": {
                        "description": "Total number of blocks (including genesis)",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/block/{height}": {
            "get": {
                "description": "Returns detailed information about a specific block at the given height. Optionally includes statistical data such as transaction count, gas usage, and fees.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Get block by height",
                "operationId": "get-block",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 12345,
                        "description": "Block height (block number)",
                        "name": "height",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include block statistics (transaction count, gas usage, fees) (default: false)",
                        "name": "stats",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Block information",
                        "schema": {
                            "$ref": "#/definitions/responses.Block"
                        }
                    },
                    "204": {
                        "description": "Block not found"
                    },
                    "400": {
                        "description": "Invalid block height",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/block/{height}/stats": {
            "get": {
                "description": "Returns statistical data for a specific block including transaction count, gas used, gas limit, base fee, and total fees",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Get block statistics",
                "operationId": "get-block-stats",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 12345,
                        "description": "Block height (block number)",
                        "name": "height",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Block statistics",
                        "schema": {
                            "$ref": "#/definitions/responses.BlockStats"
                        }
                    },
                    "400": {
                        "description": "Invalid block height",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/block/{height}/transactions": {
            "get": {
                "description": "Returns a paginated list of all transactions included in the specified block",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "List transactions in block",
                "operationId": "list-block-transactions",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 12345,
                        "description": "Block height (block number)",
                        "name": "height",
                        "in": "path",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of transactions to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of transactions to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Sort order by index (default: asc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions in the block",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid block height",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.Block": {
            "description": "Noble block information",
            "type": "object",
            "properties": {
                "difficulty": {
                    "type": "string",
                    "example": "0x0"
                },
                "extra_data": {
                    "type": "string",
                    "example": "0x726574682f76312e372e302f6c696e7578"
                },
                "gas_limit": {
                    "type": "integer",
                    "example": 1000000
                },
                "gas_used": {
                    "type": "integer",
                    "example": 500000
                },
                "hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22c0ee59781e54858a7d"
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "logs_bloom": {
                    "type": "string",
                    "example": "0x0000000000000000000020000000000"
                },
                "miner": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000000"
                },
                "mix_hash": {
                    "type": "string",
                    "example": "0x000000000000000000000000000000000000000000000000000000000033a87e"
                },
                "nonce": {
                    "type": "integer",
                    "example": 0
                },
                "parent_hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22c0ee59781e54858a7d"
                },
                "receipts_root": {
                    "type": "string",
                    "example": "0x24e9aae3033f9ff809675831eca331b701440009592a40a6d788756f3be983a2"
                },
                "sha3_uncles_hash": {
                    "type": "string",
                    "example": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347"
                },
                "size": {
                    "type": "integer",
                    "example": 0
                },
                "state_root": {
                    "type": "string",
                    "example": "0x9b6e76e8263c5060b61e396c65baf15dd187386d5607250be0dcc5308f0b49ef"
                },
                "stats": {
                    "$ref": "#/definitions/responses.BlockStats"
                },
                "time": {
                    "type": "string",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "transactions_root_hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22c0ee59781e54858a7d"
                }
            }
        },
        "responses.BlockStats": {
            "description": "Block statistics information",
            "type": "object",
            "properties": {
                "block_time": {
                    "type": "integer",
                    "example": 1000
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "time": {
                    "type": "string",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "tx_count": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "responses.Transaction": {
            "description": "Noble transaction information",
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000000000000000000
                },
                "cumulative_gas_used": {
                    "type": "integer",
                    "example": 21000
                },
                "effective_gas_price": {
                    "type": "integer",
                    "example": 1000000
                },
                "fee": {
                    "type": "string",
                    "example": "21000000000"
                },
                "from_address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000000"
                },
                "gas": {
                    "type": "integer",
                    "example": 21000
                },
                "gas_price": {
                    "type": "integer",
                    "example": 1000000
                },
                "gas_used": {
                    "type": "integer",
                    "example": 21000
                },
                "hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22c0ee59781e54858a7d"
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "index": {
                    "type": "integer",
                    "example": 0
                },
                "input": {
                    "type": "string",
                    "example": "0x"
                },
                "logs_bloom": {
                    "type": "string",
                    "example": "0x00000000000000000000000000000000000000000000"
                },
                "nonce": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "TxStatusSuccess",
                        "TxStatusRevert"
                    ],
                    "example": "TxStatusSuccess"
                },
                "time": {
                    "type": "string",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "to_address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000001"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "TxTypeUnknown",
                        "TxTypeLegacy",
                        "TxTypeDynamicFee",
                        "TxTypeBlob",
                        "TxTypeSetCode"
                    ],
                    "example": "TxTypeDynamicFee"
                }
            }
        },
...
    }
}